feat: use u16 for triggers to support future controllers

+++ crates/hidra-ipc/src/lib.rs
@@ -17,8 +17,8 @@ pub struct PadState {
     pub ly: i16,
     pub rx: i16,
     pub ry: i16,
-    pub lt: u8,
-    pub rt: u8,
+    pub lt: u16,
+    pub rt: u16,
     //TODO: add touchpad, motion, battery, etc
 }
